Login and Create User
Made some html forms to display login info.
If you sumbit, it runs php to welcome you.
There's a button to take you to a different page to create a user. When
you make a new user, it runs some php to check if the given password and
confirm passwords match. If they do, it tells you they match. If they
don't, it says they don't and it reloads tbe create user page.

Wrote some code that should check the user table on the db to see if the
desired username exists, but the database isn't buillt yet.