Create Chatrooms and Send Messages
After you log in, you get to select which chatroom to join.
The chatroom list is populated from the database from chatroom table.
You can also create a new chatroom from this menu. It will create a table
for the messages in that room as well as add the chatroom name to chatroom
table so it can be found in the dropdown list.

Once you select a chatroom, it takes you to the message interface. It
shows all the previous messages and who sent them, ordered by time sent.
You can send your own message, but then you have to refresh the browser
manually to see the new messages (including your own)

You username and password now get checked to see if they match what is in
the database. But passwords are still just stored as text

Added a session to the login so that when you go to send messages, your id
is remembered by the browser.